requirements.txt
Flask==1.1.1
pylint
pandas
newsapi-python